Outsources gradient calculation to own class
#2783
