+++ visualization/app/codeCharta/util/algorithm/treeMapLayout/treeMapHelper.ts
@@ -1,11 +1,10 @@
-import { ColorRange, MapColors } from "./../../../codeCharta.model"
 import { getMapResolutionScaleFactor, getMarkingColor, isLeaf } from "../../codeMapHelper"
 import { CodeMapNode, ColorMode, Node, State } from "../../../codeCharta.model"
-import { Color, Vector3 } from "three"
+import { Vector3 } from "three"
 import { CodeMapBuilding } from "../../../ui/codeMap/rendering/codeMapBuilding"
 import { HierarchyRectangularNode } from "d3-hierarchy"
-import { ColorConverter } from "../../color/colorConverter"
 import { searchedNodePathsSelector } from "../../../state/selectors/searchedNodes/searchedNodePaths.selector"
+import { gradientCalculator } from "../../color/gradientCalculator"
 
 const FOLDER_HEIGHT = 2
 const MIN_BUILDING_HEIGHT = 2
@@ -222,87 +221,14 @@ export function getBuildingColor(node: CodeMapNode, { appSettings, dynamicSettin
 	}
 
 	if (colorMode === ColorMode.trueGradient) {
-		return calculateTrueGradient(mapColors, colorRange, metricValue)
+		return gradientCalculator.getColorByTrueGradient(mapColors, colorRange, metricValue)
 	}
 
 	if (colorMode === ColorMode.focusedGradient) {
-		return calculateFocusedGradient(mapColors, colorRange, metricValue)
+		return gradientCalculator.getColorByFocusedGradient(mapColors, colorRange, metricValue)
 	}
 
-	return calculateWeightedGradient(mapColors, colorRange, metricValue)
-}
-
-function calculateTrueGradient(mapColors: MapColors, colorRange: ColorRange, metricValue: number) {
-	const middle = (colorRange.from + colorRange.to) / 2
-	const neutralColorRGB = ColorConverter.convertHexToColorObject(mapColors.neutral)
-
-	if (metricValue <= middle) {
-		const neutralFactor = metricValue / middle
-		const positiveColorRGB = ColorConverter.convertHexToColorObject(mapColors.positive)
-		return ColorConverter.convertColorToHex(new Color().lerpColors(positiveColorRGB, neutralColorRGB, neutralFactor))
-	}
-
-	const negativeFactor = (metricValue - middle) / (colorRange.max - middle)
-	const negativeColorRGB = ColorConverter.convertHexToColorObject(mapColors.negative)
-	return ColorConverter.convertColorToHex(new Color().lerpColors(neutralColorRGB, negativeColorRGB, negativeFactor))
-}
-
-function calculateFocusedGradient(mapColors: MapColors, colorRange: ColorRange, metricValue: number) {
-	const middle = (colorRange.from + colorRange.to) / 2
-	const neutralColorRGB = ColorConverter.convertHexToColorObject(mapColors.neutral)
-
-	if (metricValue <= colorRange.from) {
-		return mapColors.positive
-	}
-
-	if (metricValue >= colorRange.to) {
-		return mapColors.negative
-	}
-
-	if (metricValue === middle) {
-		return mapColors.neutral
-	}
-
-	if (metricValue <= middle) {
-		const neutralFactor = metricValue / (middle + colorRange.from)
-		const positiveColorRGB = ColorConverter.convertHexToColorObject(mapColors.positive)
-		return ColorConverter.convertColorToHex(new Color().lerpColors(positiveColorRGB, neutralColorRGB, neutralFactor))
-	}
-
-	const negativeFactor = (metricValue - middle) / (colorRange.to - middle)
-	const negativeColorRGB = ColorConverter.convertHexToColorObject(mapColors.negative)
-	return ColorConverter.convertColorToHex(new Color().lerpColors(neutralColorRGB, negativeColorRGB, negativeFactor))
-}
-
-function calculateWeightedGradient(mapColors: MapColors, colorRange: ColorRange, metricValue: number) {
-	const endPositive = Math.max(colorRange.from - (colorRange.to - colorRange.from) / 2, colorRange.from / 2)
-	const startNeutral = 2 * colorRange.from - endPositive
-	const endNeutral = colorRange.to - (colorRange.to - colorRange.from) / 2
-	const startNegative = colorRange.to
-
-	if (metricValue <= endPositive) {
-		return mapColors.positive
-	}
-
-	if (metricValue < startNeutral) {
-		const factor = (metricValue - endPositive) / (startNeutral - endPositive)
-		const positiveColorRGB = ColorConverter.convertHexToColorObject(mapColors.positive)
-		const neutralColorRGB = ColorConverter.convertHexToColorObject(mapColors.neutral)
-		return ColorConverter.convertColorToHex(new Color().lerpColors(positiveColorRGB, neutralColorRGB, factor))
-	}
-
-	if (metricValue <= endNeutral) {
-		return mapColors.neutral
-	}
-
-	if (metricValue < startNegative) {
-		const factor = (metricValue - endNeutral) / (startNegative - endNeutral)
-		const neutralColorRGB = ColorConverter.convertHexToColorObject(mapColors.neutral)
-		const negativeColorRGB = ColorConverter.convertHexToColorObject(mapColors.negative)
-		return ColorConverter.convertColorToHex(new Color().lerpColors(neutralColorRGB, negativeColorRGB, factor))
-	}
-
-	return mapColors.negative
+	return gradientCalculator.getColorByWeightedGradient(mapColors, colorRange, metricValue)
 }
 
 export const TreeMapHelper = {
